+++ browser-detection.js
@@ -6,6 +6,7 @@ var isIe9 = require('./is-ie9');
 var isIe10 = require('./is-ie10');
 var isIos = require('./is-ios');
 var isIosSafari = require('./is-ios-safari');
+var isIosWebview = require('./is-ios-webview');
 var supportsPopups = require('./supports-popups');
 
 module.exports = {
@@ -15,5 +16,6 @@ module.exports = {
   isIe10: isIe10,
   isIos: isIos,
   isIosSafari: isIosSafari,
+  isIosWebview: isIosWebview,
   supportsPopups: supportsPopups
 };