+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Browser Detection - Release Notes
 
-## unreleased
+## UNRELEASED
 
+* Add `is-ios-webview` method
 * Add `is-ios-safari` method
 
 ## 1.2.0 (2017-04-10)